♻️ Todo管理ロジックをApp.texで行うようにリファクタリング
